backend/requirements.txt
django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-cors-headers==3.14.0
channels==4.0.0
asgiref==3.6.0
daphne==4.0.0